とりあえず3Dサウンド入った

+++ gctp/include/gctp/def.hpp
@@ -38,8 +38,8 @@ namespace gctp {
 	typedef unsigned char		uint8_t;
 	typedef short				int16_t;
 	typedef unsigned short		uint16_t;
-	typedef int					int32_t;
-	typedef unsigned int		uint32_t;
+	typedef long				int32_t;
+	typedef unsigned long		uint32_t;
 	typedef __int64				int64_t;
 	typedef unsigned __int64	uint64_t;
 #endif
